Remove unnecessary integer boxing in AInt*
Reviewed-on: https://asterix-gerrit.ics.uci.edu/1797
Sonar-Qube: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
BAD: Jenkins <jenkins@fulliautomatix.ics.uci.edu>
Integration-Tests: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ asterixdb/asterix-external-data/src/main/java/org/apache/asterix/external/library/java/JObjects.java
@@ -266,7 +266,7 @@ public class JObjects {
         }
 
         public int getValue() {
-            return ((AMutableInt32) value).getIntegerValue().intValue();
+            return ((AMutableInt32) value).getIntegerValue();
         }
 
         @Override
